Download Human DLPFC Deconvolution Data

Download Human DLPFC Deconvolution Data

fetch_deconvo_data(
  type = c("rse_gene"),
  destdir = tempdir(),
  eh = ExperimentHub::ExperimentHub(),
  bfc = BiocFileCache::BiocFileCache()
)

Arguments

type

A character(1) specifying which file you want to download. rse_gene RangedSummarizedExperiment with 110 bulk RNA-seq samples

destdir

The destination directory to where files will be downloaded to in case the ExperimentHub resource is not available. If you already downloaded the files, you can set this to the current path where the files were previously downloaded to avoid re-downloading them.

eh

An ExperimentHub object ExperimentHub-class.

bfc

A BiocFileCache object BiocFileCache-class. Used when eh is not available.

Value

The requested object

Examples

rse_gene <- fetch_deconvo_data("rse_gene")
#> 2024-03-27 12:01:58.109132 loading file /github/home/.cache/R/BiocFileCache/66d183cf17f_rse_gene.Rdata%3Frlkey%3Dsw2djr71y954yw4o3xrmjv59b%26dl%3D1