Calculate pw standard fold change for each gene x cell type

Calculate pw standard fold change for each gene x cell type

findMarkers_pw(
  sce,
  assay_name = "counts",
  cellType_col = "cellType",
  add_symbol = FALSE
)

Arguments

sce

single cell experiment object

assay_name

Name of the assay to use for calculation

cellType_col

Column name on colData of the sce that denotes the celltype

add_symbol

Add the gene symbol column to the marker stats table

Value

Table of 1 vs. ALL std log fold change + p-values for each gene x cell type

Examples

markers_pw <- findMarkers_pw(sce.test)
head(markers_pw)
#> DataFrame with 6 rows and 5 columns
#>              Gene cellType.target summary.logFC     p.value         FDR
#>       <character>     <character>     <numeric>   <numeric>   <numeric>
#> 1 ENSG00000142611           Astro      2.676049 4.72522e-32 4.72522e-29
#> 2 ENSG00000117525           Astro      1.864641 1.34495e-17 6.72475e-15
#> 3 ENSG00000177133           Astro      1.457539 1.01063e-11 3.36877e-09
#> 4 ENSG00000231252           Astro      1.089676 7.06075e-10 1.76519e-07
#> 5 ENSG00000079335           Astro      0.988865 1.77609e-08 3.55217e-06
#> 6 ENSG00000273136           Astro      1.054076 4.76259e-07 7.93765e-05