Get Mean Ratio for Each Gene x Cell Type

Calculate the mean ratio value and rank for each gene for each cell type in the sce object, to identify effective marker genes for deconvolution.

get_mean_ratio(
  sce,
  cellType_col,
  assay_name = "logcounts",
  gene_ensembl = NULL,
  gene_name = NULL
)

Arguments

sce

SummarizedExperiment-class object

cellType_col

A character(1) name of the column in the colData() of sce that denotes the cell type or group of interest

assay_name

A character(1) specifying the name of the assay() in the sce object to use to rank expression values. Defaults to logcounts since it typically contains the normalized expression values.

gene_ensembl

A character(1) specifying the rowData(sce_pseudo) column with the ENSEMBL gene IDs. This will be used by layer_stat_cor().

gene_name

A character(1) specifying the rowData(sce_pseudo) column with the gene names (symbols).

Value

A tibble with the MeanRatio values for each gene x cell type.

gene is the name of the gene (from rownames(sce)).

cellType.target is the cell type we're finding marker genes for.

mean.target is the mean expression of gene for cellType.target.

cellType.2nd is the second highest non-target cell type.

mean.2nd is the mean expression of gene for cellType.2nd.

MeanRatio is the ratio of mean.target/mean.2nd.

MeanRatio.rank is the rank of MeanRatio for the cell type.

MeanRatio.anno is an annotation of the MeanRatio calculation helpful for plotting.

gene_ensembl & gene_name optional cols spcified by the user to add gene infomation

Details

Improved argument names and documentaion, but same functionalty from get_mean_ratio2().

Examples

## Get the mean ratio for each gene for each cell type defined in `cellType`
get_mean_ratio(sce.test, cellType_col = "cellType")
#> # A tibble: 1,778 × 8
#>    gene            cellType.target mean.target cellType.2nd mean.2nd MeanRatio
#>    <chr>           <fct>                 <dbl> <fct>           <dbl>     <dbl>
#>  1 ENSG00000230615 Inhib.2               1.00  Excit.2         0.239      4.20
#>  2 ENSG00000162512 Inhib.2               1.71  Astro           0.512      3.35
#>  3 ENSG00000137965 Inhib.2               0.950 Astro           0.413      2.30
#>  4 ENSG00000060718 Inhib.2               3.32  Astro           1.47       2.26
#>  5 ENSG00000162631 Inhib.2               3.55  Astro           1.62       2.19
#>  6 ENSG00000116525 Inhib.2               1.22  Excit.1         0.560      2.18
#>  7 ENSG00000162402 Inhib.2               3.37  Excit.2         1.96       1.72
#>  8 ENSG00000132122 Inhib.2               3.42  Inhib.1         2.44       1.40
#>  9 ENSG00000117528 Inhib.2               1.21  Astro           0.914      1.33
#> 10 ENSG00000172380 Inhib.2               1.21  Astro           0.920      1.32
#> # ℹ 1,768 more rows
#> # ℹ 2 more variables: MeanRatio.rank <int>, MeanRatio.anno <chr>

#   gene            cellType.target mean.target cellType.2nd mean.2nd MeanRatio MeanRatio.rank MeanRatio.anno        
#  <chr>           <fct>                 <dbl> <fct>           <dbl>     <dbl>          <int> <chr>                 
# 1 ENSG00000230615 Inhib.2               1.00  Excit.2         0.239      4.20              1 Inhib.2/Excit.2: 4.197
# 2 ENSG00000162512 Inhib.2               1.71  Astro           0.512      3.35              2 Inhib.2/Astro: 3.346  
# 3 ENSG00000137965 Inhib.2               0.950 Astro           0.413      2.30              3 Inhib.2/Astro: 2.301  
# 4 ENSG00000060718 Inhib.2               3.32  Astro           1.47       2.26              4 Inhib.2/Astro: 2.264  
# 5 ENSG00000162631 Inhib.2               3.55  Astro           1.62       2.19              5 Inhib.2/Astro: 2.19  

# Option to specify gene_name as the "Symbol" column from rowData
# this will be added to the marker stats output
SummarizedExperiment::rowData(sce.test)
#> DataFrame with 1000 rows and 3 columns
#>                         gene_id bp_length      Symbol
#>                     <character> <integer> <character>
#> ENSG00000237491 ENSG00000237491     31291  AL669831.5
#> ENSG00000225880 ENSG00000225880      1317   LINC00115
#> ENSG00000187634 ENSG00000187634     20654      SAMD11
#> ENSG00000188976 ENSG00000188976     15106       NOC2L
#> ENSG00000187961 ENSG00000187961      5129      KLHL17
#> ...                         ...       ...         ...
#> ENSG00000183598 ENSG00000183598       411    HIST2H3D
#> ENSG00000272993 ENSG00000272993       671  AC239868.3
#> ENSG00000184678 ENSG00000184678      2194   HIST2H2BE
#> ENSG00000184260 ENSG00000184260       390   HIST2H2AC
#> ENSG00000184270 ENSG00000184270       393   HIST2H2AB
get_mean_ratio(sce.test, cellType_col = "cellType", gene_name = "Symbol", gene_ensembl = "gene_id")
#> # A tibble: 1,778 × 10
#>    gene            cellType.target mean.target cellType.2nd mean.2nd MeanRatio
#>    <chr>           <fct>                 <dbl> <fct>           <dbl>     <dbl>
#>  1 ENSG00000230615 Inhib.2               1.00  Excit.2         0.239      4.20
#>  2 ENSG00000162512 Inhib.2               1.71  Astro           0.512      3.35
#>  3 ENSG00000137965 Inhib.2               0.950 Astro           0.413      2.30
#>  4 ENSG00000060718 Inhib.2               3.32  Astro           1.47       2.26
#>  5 ENSG00000162631 Inhib.2               3.55  Astro           1.62       2.19
#>  6 ENSG00000116525 Inhib.2               1.22  Excit.1         0.560      2.18
#>  7 ENSG00000162402 Inhib.2               3.37  Excit.2         1.96       1.72
#>  8 ENSG00000132122 Inhib.2               3.42  Inhib.1         2.44       1.40
#>  9 ENSG00000117528 Inhib.2               1.21  Astro           0.914      1.33
#> 10 ENSG00000172380 Inhib.2               1.21  Astro           0.920      1.32
#> # ℹ 1,768 more rows
#> # ℹ 4 more variables: MeanRatio.rank <int>, MeanRatio.anno <chr>,
#> #   gene_ensembl <chr>, gene_name <chr>
# A tibble: 1,778 × 10
# gene            cellType.target mean.target cellType.2nd mean.2nd MeanRatio MeanRatio.rank MeanRatio.anno   gene_ensembl gene_name
# <chr>           <fct>                 <dbl> <fct>           <dbl>     <dbl>          <int> <chr>            <chr>        <chr>    
# 1 ENSG00000230615 Inhib.2               1.00  Excit.2         0.239      4.20              1 Inhib.2/Excit.2… ENSG0000023… AL139220…
# 2 ENSG00000162512 Inhib.2               1.71  Astro           0.512      3.35              2 Inhib.2/Astro: … ENSG0000016… SDC3     
# 3 ENSG00000137965 Inhib.2               0.950 Astro           0.413      2.30              3 Inhib.2/Astro: … ENSG0000013… IFI44