Get Mean Ratio for Each Gene x Cell Type

Get Mean Ratio for Each Gene x Cell Type

get_mean_ratio2(
  sce,
  cellType_col = "cellType",
  assay_name = "logcounts",
  add_symbol = TRUE
)

Arguments

sce
Single cell experiment object

cellType_col
Column name on colData of the sce that denotes the celltype

assay_name
Name of assay to use for calculation

Value

Table of mean ratio for each x cell type

Examples

get_mean_ratio2(sce.test)

#> # A tibble: 2,504 x 9
#>    gene  cellType.target mean.target cellType  mean ratio rank_ratio Symbol
#>    <chr> <fct>                 <dbl> <fct>    <dbl> <dbl>      <int> <chr> 
#>  1 ENSG… Oligo                  5.98 Astro    1.29   4.62          1 RNF220
#>  2 ENSG… Oligo                  2.19 OPC      0.737  2.98          2 CLIC4 
#>  3 ENSG… Oligo                  2.73 Micro    1.35   2.02          3 ATG4C 
#>  4 ENSG… Oligo                  3.36 Excit.3  1.70   1.98          4 SORT1 
#>  5 ENSG… Oligo                  2.83 Excit.4  1.50   1.88          5 SLC22…
#>  6 ENSG… Oligo                  5.16 Excit.1  2.90   1.78          6 TTLL7 
#>  7 ENSG… Oligo                  3.68 OPC      2.26   1.63          7 PTBP2 
#>  8 ENSG… Oligo                  1.52 Astro    0.944  1.61          8 PADI2 
#>  9 ENSG… Oligo                  1.57 Micro    1.01   1.55          9 SRRM1 
#> 10 ENSG… Oligo                  4.11 Excit.2  2.67   1.54         10 DNAJC6
#> # … with 2,494 more rows, and 1 more variable: anno_ratio <chr>