Plot the top marker genes for ALL cell types

This function plots the top n marker genes for a all cell types based off of the stats table from get_mean_ratio2() in a multi-page pdf. The gene expression is plotted as violin plot with plot_gene_express and adds annotations to each plot.

plot_marker_express_ALL(
  sce,
  stats,
  pdf_fn = "marker_expression.pdf",
  n_genes = 10,
  rank_col = "rank_ratio",
  anno_col = "anno_ratio",
  cellType_col = "cellType",
  color_pal = NULL,
  plot_points = FALSE
)

Arguments

sce

SummarizedExperiment-class object

stats

A data.frame() generated by get_mean_ratio and/or findMarkers_1vAll

pdf_fn

A character() of the pdf filename to plot to

n_genes

An integer() of number of markers you'd like to plot

rank_col

The character() name of column to rank genes by in stats

anno_col

The character() name of column containing annotation in stats

cellType_col

The character() name of colData column containing cell type for sce data, matches cellType.target in stats

color_pal

A named character(1) vector that contains a color pallet matching the values in cellType_col.

plot_points

A logical indicating whether to plot points over the violin, defaults to FALSE as these often become over plotted and quite large (especially when saved as PDF)

Value

A pdf with violin plots for the expression of top marker genes for all cell types

See also

Examples

# plot_marker_express_ALL(sce.test, stat = marker_test, pdf_fn = "./plots/test_marker_expression_ALL.pdf")