Calculate 1 vs. All standard fold change for each gene x cell type, wrapper function for scran::findMarkers

Calculate 1 vs. All standard fold change for each gene x cell type, wrapper function for scran::findMarkers

findMarkers_1vAll(
  sce,
  assay_name = "counts",
  cellType_col = "cellType",
  add_symbol = FALSE,
  mod = "~donor",
  verbose = TRUE
)

Arguments

sce

single cell experiment object

assay_name

Name of the assay to use for calculation

cellType_col

Column name on colData of the sce that denotes the celltype

add_symbol

Add the gene symbol column to the marker stats table

mod

String specifying the model used as design in findMarkers. Can be NULL if there are no blocking terms with uninteresting factors as documented at pairwiseTTests.

verbose

Boolean choosing to print progress messages or not

Value

Table of 1 vs. ALL std log fold change + p-values for each gene x cell type

Examples

markers_1vAll <- findMarkers_1vAll(sce.test)
#> Inhib.2 - '2024-05-02 22:22:27.594127
#> Inhib.1 - '2024-05-02 22:22:27.92237
#> OPC - '2024-05-02 22:22:28.837038
#> Astro - '2024-05-02 22:22:28.999362
#> Excit.2 - '2024-05-02 22:22:29.163615
#> Oligo - '2024-05-02 22:22:29.340659
#> Micro - '2024-05-02 22:22:29.516513
#> Excit.1 - '2024-05-02 22:22:29.678743
#> Building Table - 2024-05-02 22:22:29.842259
#> ** Done! **
head(markers_1vAll)
#> # A tibble: 6 × 8
#> # Groups:   cellType.target [1]
#>   gene           logFC log.p.value log.FDR std.logFC cellType.target rank_marker
#>   <chr>          <dbl>       <dbl>   <dbl>     <dbl> <fct>                 <int>
#> 1 ENSG000001321… 5.33        -83.1   -76.2      1.96 Inhib.2                   1
#> 2 ENSG000002306… 0.928       -63.6   -57.3      1.67 Inhib.2                   2
#> 3 ENSG000001379… 0.753       -48.4   -42.6      1.42 Inhib.2                   3
#> 4 ENSG000000490… 0.789       -41.3   -35.8      1.30 Inhib.2                   4
#> 5 ENSG000001625… 1.27        -38.2   -32.9      1.24 Inhib.2                   5
#> 6 ENSG000001219… 1.60        -37.7   -32.6      1.23 Inhib.2                   6
#> # ℹ 1 more variable: anno_logFC <chr>