Get Mean Ratio for Each Gene x Cell Type

Get Mean Ratio for Each Gene x Cell Type

get_mean_ratio(
  sce,
  cellType_col = "cellType",
  assay_name = "logcounts",
  add_symbol = FALSE
)

Arguments

sce

Single cell experiment object

cellType_col

Column name on colData of the sce that denotes the celltype

assay_name

Name of the assay to use for calculation

add_symbol

Add the gene symbol column to the marker stats table

Value

Table of mean ratio for each x cell type

Examples

markers_ratio <- get_mean_ratio(sce.test)
#> nrow cellType: 500
#> build target stat
#> build mean ratio
head(markers_ratio)
#> # A tibble: 6 × 10
#> # Groups:   cellType.target [1]
#>   gene   cellType.target median.target mean.target cellType median    mean ratio
#>   <fct>  <fct>                   <dbl>       <dbl> <fct>     <dbl>   <dbl> <dbl>
#> 1 ENSG0… Excit.2                 2.55         2.11 Astro         0 0       Inf  
#> 2 ENSG0… Excit.2                 1.57         1.38 Oligo         0 0.00763 181. 
#> 3 ENSG0… Excit.2                 0.820        1.30 Astro         0 0.0138   94.9
#> 4 ENSG0… Excit.2                 2.10         1.63 OPC           0 0.0438   37.3
#> 5 ENSG0… Excit.2                 1.89         1.61 Micro         0 0.0590   27.3
#> 6 ENSG0… Excit.2                 1.98         1.81 Inhib.2       0 0.0907   20.0
#> # ℹ 2 more variables: rank_ratio <int>, anno_ratio <chr>