Get Mean Ratio for Each Gene x Cell Type

Get Mean Ratio for Each Gene x Cell Type

get_mean_ratio2(
  sce,
  cellType_col = "cellType",
  assay_name = "logcounts",
  add_symbol = TRUE
)

Arguments

sce
Single cell experiment object

cellType_col
Column name on colData of the sce that denotes the celltype

assay_name
Name of assay to use for calculation

Value

Table of mean ratio for each x cell type

Examples

get_mean_ratio2(sce.test)

#> # A tibble: 1,750 x 9
#>    gene  cellType.target mean.target cellType   mean ratio rank_ratio Symbol
#>    <chr> <fct>                 <dbl> <fct>     <dbl> <dbl>      <int> <chr> 
#>  1 ENSG… Excit.1               0.508 Inhib.2  0.0260 19.5           1 AL136…
#>  2 ENSG… Excit.1               0.718 Inhib.2  0.0448 16.0           2 PALMD 
#>  3 ENSG… Excit.1               0.769 Excit.2  0.0992  7.75          3 PTGER3
#>  4 ENSG… Excit.1               1.73  Excit.2  0.656   2.64          4 COL24…
#>  5 ENSG… Excit.1               0.599 Excit.2  0.266   2.25          5 LINC0…
#>  6 ENSG… Excit.1               0.433 Excit.2  0.239   1.81          6 ZNF697
#>  7 ENSG… Excit.1               0.517 Astro    0.295   1.75          7 AL139…
#>  8 ENSG… Excit.1               2.18  Inhib.1  1.27    1.72          8 MAN1C1
#>  9 ENSG… Excit.1               0.265 Excit.2  0.157   1.69          9 AL591…
#> 10 ENSG… Excit.1               0.856 Excit.2  0.512   1.67         10 TNFRS…
#> # … with 1,740 more rows, and 1 more variable: anno_ratio <chr>